+++ packages/terra-clinical-detail-view/src/terra-dev-site/test/clinical-detail-view/DetailView/DetailViewDividedSmallerTitles.test.jsx
@@ -28,7 +28,7 @@ const DetailViewDividedSmallerTitles = () => (
       <div className={cx('graph-content')}>
         This is where any visualizations would go
       </div>
-)}
+    )}
     accessory={accessoryElement}
     details={[
       (