+++ packages/terra-clinical-result/CHANGELOG.md
@@ -3,9 +3,15 @@
 ## Unreleased
 
 * Added
+  * Added screen-reader support for strikethroughs to convey an entered in error status.
+
+## 1.17.0 - (June 14, 2023)
+
+* Added
   * Added screen-reader support for clinical-result icons. (Requires Jest test updates on consuming applications)
   * Added screen-reader support for alternative text for when no results are returned.
   * Added screen-reader support for clinical-result FlowsheetResultCell EnteredInError.
+  * Added screen-reader support for clinical-result FlowsheetResultCell with multiple results.
   * Added screen-reader support for clinical-result ResultNameHeaderCell. Two new optional props are added for providing a full name of the result name and unit respectively.
 
 ## 1.16.0 - (March 29, 2023)